+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,62 +1,6 @@
 import NextAuth from 'next-auth';
-import CredentialsProvider from 'next-auth/providers/credentials';
-
-export const authOptions = {
-  providers: [
-    CredentialsProvider({
-      name: 'Credentials',
-      credentials: {
-        password: { label: 'Password', type: 'password' },
-        isBackoffice: { label: 'Is Backoffice', type: 'boolean' }
-      },
-      async authorize(credentials) {
-        if (!credentials?.password) {
-          return null;
-        }
-
-        try {
-          const response = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/auth/${credentials.isBackoffice ? 'backoffice' : 'public'}/login`, {
-            method: 'POST',
-            headers: {
-              'Content-Type': 'application/json',
-            },
-            body: JSON.stringify({
-              password: credentials.password,
-            }),
-          });
-
-          if (!response.ok) {
-            return null;
-          }
-
-          const data = await response.json();
-          return {
-            id: '1',
-            accessToken: data.access_token,
-          };
-        } catch (error) {
-          console.error('Auth error:', error);
-          return null;
-        }
-      },
-    }),
-  ],
-  callbacks: {
-    async jwt({ token, user }) {
-      if (user) {
-        token.accessToken = user.accessToken;
-      }
-      return token;
-    },
-    async session({ session, token }) {
-      session.accessToken = token.accessToken;
-      return session;
-    },
-  },
-  pages: {
-    signIn: '/login',
-  },
-};
+import { authOptions } from '@/lib/auth';
 
 const handler = NextAuth(authOptions);
+
 export { handler as GET, handler as POST };
